requirements/base.txt
certifi==2017.11.5
chardet==3.0.4
Django==1.11.7
djangorestframework==3.7.3
httpie==0.9.9
idna==2.6
olefile==0.44
Pillow==4.3.0
Pygments==2.2.0
pytz==2017.3
requests==2.18.4
urllib3==1.22
